fix: lint code and refactor namespace

+++ packages/libs/common/wraptypes/basic-types.ts
@@ -1,4 +1,3 @@
-
 export enum typeIndex {
   Empty = 0,
   TypeInt8,
@@ -57,7 +56,7 @@ export class AString {
   }
 
   public store(): i32 {
-    const ptr = heap.alloc(this.head.dataLen + this.head.len());
+    const ptr = heap.alloc(this.head.dataLen + this.head.len()) as i32;
     this.head.store(ptr);
     const bodyPtr = ptr + this.head.len();
     // utf-16 <--> utf8
@@ -99,7 +98,7 @@ export class AUint8Array {
   }
 
   public store(): i32 {
-    const ptr = heap.alloc(this.head.dataLen + this.head.len());
+    const ptr = heap.alloc(this.head.dataLen + this.head.len()) as i32;
     this.head.store(ptr);
     let bodyPtr = ptr + this.head.len();
     for (let i = 0; i < this.head.dataLen; i++) {
@@ -136,7 +135,7 @@ export class ABool {
   }
 
   public store(): i32 {
-    const ptr = heap.alloc(this.head.dataLen + this.head.len());
+    const ptr = heap.alloc(this.head.dataLen + this.head.len()) as i32;
     this.head.store(ptr);
     const bodyPtr = ptr + this.head.len();
     memory.fill(bodyPtr, this.body ? 1 : 0, 1);
@@ -170,7 +169,7 @@ export class AI32 {
   }
 
   public store(): i32 {
-    const ptr = heap.alloc(this.head.dataLen + this.head.len());
+    const ptr = heap.alloc(this.head.dataLen + this.head.len()) as i32;
     this.head.store(ptr);
     const bodyPtr = ptr + this.head.len();
     i32.store(bodyPtr, this.body);
@@ -204,7 +203,7 @@ export class AI64 {
   }
 
   public store(): i64 {
-    const ptr = heap.alloc(this.head.dataLen + this.head.len());
+    const ptr = heap.alloc(this.head.dataLen + this.head.len()) as i32;
     this.head.store(ptr);
     const bodyPtr = ptr + this.head.len();
     i64.store(bodyPtr, this.body);
